requirements.txt
###pip install -i https://pypi.doubanio.com/simple/ xxx ###

django
djangorestframework
django-filter
markdown
mysqlclient
coreapi
djangorestframework-jwt

numpy
gensim
StanfordCoreNLP
pymysql
scipy
zhon
jieba
nltk
joblib
scikit-learn